static_cast & remove FrameBuffer glEnable depth test

+++ OGL/src/Program.cpp
@@ -83,9 +83,9 @@ GLuint Program::getHandler() const
     return this->handler;
 }
 
-void Program::applyCompute(glm::ivec3 groups)
+void Program::applyCompute(glm::uvec3 groups)
 {
-    glDispatchCompute((GLuint)groups.x, (GLuint)groups.y, (GLuint)groups.z);
+    glDispatchCompute(groups.x, groups.y, groups.z);
     glMemoryBarrier(GL_SHADER_IMAGE_ACCESS_BARRIER_BIT);
 }
